reorganization and bug fixes for volunteer management tab
